feat: new layout WiP
