+++ src/components/Layers.module.css
@@ -1,10 +1,8 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 
-.palette {
-  display: flex;
-  flex-wrap: wrap;
-  margin: 0 auto;
+.layersPanel {
   border: solid #aaa 1px;
+  margin-right: 12px;
   padding: 12px 12px;
   border-radius: 12px;
 }
